Caching.ExpiringCacheMap.Utils.Types

Description

Simple types.

Types

type TimeUnits = Int

Integer involved in the time units used to determine when an item expires. The time units used can be any arbitrary integer time representation, such as seconds or milliseconds for examples. They can also be deterministic time steps in a sequencing monad.

type ECMMapSize = Int

Integer involved in the size of a key-value map.

type ECMULength = Int

Integer involved in the length of the usage history list.

type ECMIncr = Word32

Unsigned integer (Word32) involved in the cache state incrementing accumulator.